+++ src/features/incoming-report/index.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const IncomingReportPage = () => {
+  return (
+    <div>IncomingReportPage</div>
+  )
+}
+
+export default IncomingReportPage